①修改roomid生成 ②房间信息存储到redis

+++ src/main/java/com/game/firstgameserver/iogame/msg/PlayroomInfo.java
@@ -34,7 +34,7 @@ public class PlayroomInfo {
     /**
      * 创建者ID
      */
-    long primaryuserID;
+    int primaryuserID;
     /**
      * 房间最大人数
      */
